requirements.txt
pandas
yfinance
python-telegram-bot==13.15
pytz
